+++ ctest_cmake.sh
@@ -1,0 +1,109 @@
+#! /bin/bash
+
+ScriptPath=$0
+Dir=$(cd $(dirname "$ScriptPath"); pwd)
+Basename=$(basename "$ScriptPath")
+CMakeDir=${SIS_CMAKE_BUILD_DIR:-$Dir/_build}
+[[ -n "$MSYSTEM" ]] && DefaultMakeCmd=mingw32-make.exe || DefaultMakeCmd=make
+MakeCmd=${SIS_CMAKE_MAKE_COMMAND:-${SIS_CMAKE_COMMAND:-$DefaultMakeCmd}}
+
+CMakeVerbose=
+RunMake=1
+
+
+# ##########################################################
+# command-line handling
+
+while [[ $# -gt 0 ]]; do
+
+  case $1 in
+    --no-make|-M)
+
+      RunMake=0
+      ;;
+    --verbose|-V)
+
+      CMakeVerbose=--verbose
+      ;;
+    --help)
+
+      cat << EOF
+shwild is a small, standalone library, implemented in C++ with a C and a C++ API, that provides shell-compatible wildcard matching
+Copyright (c) 2019-2025, Matthew Wilson and Synesis Information Systems
+Copyright (c) 2005-2023, Matthew Wilson and Sean Kelly
+Runs CMake's CTest test program(s)
+
+$ScriptPath [ ... flags/options ... ]
+
+Flags/options:
+
+    behaviour:
+
+    -M
+    --no-make
+        does not execute CMake and make before running tests
+
+    -V
+    --verbose
+        verbose test output
+
+
+    standard flags:
+
+    --help
+        displays this help and terminates
+
+EOF
+
+      exit 0
+      ;;
+    *)
+
+      >&2 echo "$ScriptPath: unrecognised argument '$1'; use --help for usage"
+
+      exit 1
+      ;;
+  esac
+
+  shift
+done
+
+
+# ##########################################################
+# main()
+
+status=0
+
+if [ $RunMake -ne 0 ]; then
+
+  echo "Executing build (via command \`$MakeCmd\`) and then running all component and unit test programs"
+
+  mkdir -p $CMakeDir || exit 1
+
+  cd $CMakeDir
+
+  $MakeCmd
+  status=$?
+
+  cd ->/dev/null
+else
+
+  if [ ! -d "$CMakeDir" ] || [ ! -f "$CMakeDir/CMakeCache.txt" ] || [ ! -d "$CMakeDir/CMakeFiles" ]; then
+
+    >&2 echo "$ScriptPath: cannot run in '--no-make' mode without a previous successful build step"
+  fi
+fi
+
+if [ $status -eq 0 ]; then
+
+  echo "Running CMake tests"
+
+  ctest --test-dir $CMakeDir $CMakeVerbose
+  status=$?
+fi
+
+exit $status
+
+
+# ############################## end of file ############################# #
+