+++ CHANGES.md
@@ -3,6 +3,8 @@
 
 ## 30th August 2025 - 0.12.7
 
+* GitHub Actions;
+* CTest compatibility;
 * updated **BDUT** to 0.4.0;
 * removed Visual Studio project/solution/workspace files;
 * refactoring CMake support files;